fix builds with new webkit version

+++ build.sh
@@ -1,8 +1,15 @@
 echo "Building node-webkit..."
 
+if [ "$KEY_PASSWORD" == "" ]; then
+  echo "KEY_PASSWORD is blank"
+  exit 1
+fi
+
 BUILD_DIR=`pwd`
 NODE_WEBKIT_VERSION=0.11.5
 ulimit -n 8192
+
+rm -rf build
 
 echo "Cloning gateblu-ui..."
 rm -rf gateblu-ui
@@ -35,8 +42,8 @@ cd build/Gateblu/win32
 mkdir -p dist/node-v0.10.32-win-x86/bin
 cp $BUILD_DIR/dist/node-v0.10.32-win-x86/node.exe dist/node-v0.10.32-win-x86/bin
 unzip -q $BUILD_DIR/dist/node-v0.10.32-win-x86/npm-1.4.9.zip -d dist/node-v0.10.32-win-x86/bin
-cp -rfp $BUILD_DIR/cache/$NODE_WEBKIT_VERSION/win/locales locales
 mv Gateblu.exe package.nw
+cp $BUILD_DIR/cache/$NODE_WEBKIT_VERSION/win32/nw.exe Gateblu.exe
 
 echo "Creating Windows Installer..."
 cd $BUILD_DIR
@@ -48,8 +55,6 @@ cd build/Gateblu/linux32
 mkdir dist
 tar zxf $BUILD_DIR/dist/node-v0.10.32-linux-x86.tar.gz --directory dist/
 LC_CTYPE=C && LANG=C && sed -i '' 's/udev\.so\.0/udev.so.1/g' Gateblu
-cp -rfp $BUILD_DIR/cache/$NODE_WEBKIT_VERSION/linux32/locales locales
-mv Gateblu package.nw
 
 echo "Building Linux 64-bit..."
 cd $BUILD_DIR
@@ -57,8 +62,6 @@ cd build/Gateblu/linux64
 mkdir dist
 tar zxf $BUILD_DIR/dist/node-v0.10.32-linux-x64.tar.gz --directory dist/
 LC_CTYPE=C && LANG=C && sed -i '' 's/udev\.so\.0/udev.so.1/g' Gateblu
-cp -rfp $BUILD_DIR/cache/$NODE_WEBKIT_VERSION/linux64/locales locales
-mv Gateblu package.nw
 
 echo "Creating Linux Installer..."
 cd $BUILD_DIR
